refactor: keep settings types type-only
